requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.6.0

# Database
psycopg2-binary>=2.9.0

# ML
lightgbm>=4.0.0
scikit-learn>=1.3.0
optuna>=3.3.0

# Dashboard
flask>=3.0.0

# Data
pyarrow>=14.0.0

# Deprecated (kept for backward compat, not used in v5.0)
# yfinance
# ccxt
# schedule
